+++ lib/plug/accesslog/formatter/request_serving_time.ex
@@ -1,6 +1,6 @@
 defmodule Plug.AccessLog.Formatter.RequestServingTime do
   @moduledoc """
-  Calculates the seconds taken to serve the request.
+  Calculates the seconds (or milliseconds) taken to serve the request.
   """
 
   use Timex
@@ -8,15 +8,15 @@ defmodule Plug.AccessLog.Formatter.RequestServingTime do
   @doc """
   Appends to log output.
   """
-  @spec append(String.t, Plug.Conn.t) :: String.t
-  def append(message, conn), do: message <> serving_time(conn)
+  @spec append(String.t, Plug.Conn.t, :secs | :msecs) :: String.t
+  def append(message, conn, format), do: message <> serving_time(conn, format)
 
-  defp serving_time(conn) do
+  defp serving_time(conn, format) do
     conn.private[:plug_accesslog]
     |> Map.get(:timestamp)
     |> Time.sub(:os.timestamp())
     |> Time.invert()
-    |> Time.convert(:secs)
+    |> Time.convert(format)
     |> round()
     |> to_string()
   end