+++ test/plug/accesslog/formatter/request_serving_time_test.exs
@@ -6,6 +6,19 @@ defmodule Plug.AccessLog.Formatter.RequestServingTimeTest do
 
   alias Plug.AccessLog.Formatter
 
+  test "%M" do
+    timestamp = :os.timestamp()
+    conn      =
+         conn(:get, "/")
+      |> put_private(:plug_accesslog, %{ timestamp: timestamp })
+
+    # relying on millisecond testing precision is flaky.
+    # check done using "milliseconds taken < 1 second"
+    { msecs, _ } = Formatter.format("%M", conn) |> Integer.parse()
+
+    assert msecs < 1000
+  end
+
   test "%T" do
     timestamp = :os.timestamp()
     conn      =